[#305] refactor market-list library
