+++ libs/market-list/src/index.ts
@@ -1,6 +1,2 @@
-export * from './lib/market-list-table';
-export * from './lib/markets-container';
-export * from './lib/landing-dialog';
-export * from './lib/__generated__/Markets';
-export * from './lib/__generated__/MarketDataFields';
-export * from './lib/__generated__/MarketDataSub';
+export * from './lib/components';
+export * from './lib/utils';
