+++ libs/market-list/src/lib/components/markets-container/summary-cell.tsx
